backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
aiofiles==23.2.1
httpx==0.25.2
python-multipart==0.0.6
pydantic>=2.8.0
asyncio-mqtt==0.16.1
openai==1.3.7
anthropic==0.7.8
requests==2.31.0
beautifulsoup4==4.12.2
dnspython==2.4.2
python-nmap==0.7.1
shodan==1.30.1
censys==2.2.8
whois==0.9.27
sqlparse==0.4.4
jinja2==3.1.2
reportlab==4.0.7
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
nltk==3.8.1
redis==5.0.1
aioredis==2.0.1
celery==5.3.4